Get metadata for images
Adds `MetadataCollector` that can get name, URL and license.

If an image is a digital representation the name is taken first from the works title and then label. If neither of them exist or if the image isn't a digital representation the filename is used.

URL points to the image page using Special:Redirect.

The license is the URL of the license from P856 ("official website"). It's first taken from a digital representation if there is any. Otherwise it's taken from the image itself.

Task: T390950