Add declaration journal (#8)
The journal is used to track what images have been processed. Skips ISCC generation for images for which that has already been done. The journal is stored as a database. Fields for journal entries include:

* timestamp - when the entry was added
* page_id
* revision_id
* image_hash - from Commons
* iscc

Adds the flag `--iscc`/`-i` to only generate ISCC for an image. This can be used to do this beforehand and used in later requests. Using ISCC from journal isn't yet implemented and will be added in a future update.

Task: T395105