Add tags to journal (#17)
A set of tags can be added to each declaration. The tags are stored in a
separate table and that is expanded when needed.

A batch tag with the label "batch:<FILENAME WITHOUT EXTENSITON>" is added to
all declarations.

Also adds an argument to quit on the first error when processing files.

Task: T396868